pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-classic>=1.0.0",
    "langchain-cohere>=0.3.11",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.0.9",
    "langchain-openai>=0.3.31",
    "langchain-pinecone>=0.2.11",
    "langgraph>=0.2.64",
    "langgraph-checkpoint>=2.0.13",
    "langgraph-sdk>=0.1.45",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.1.1",
    "streamlit>=1.40.0",
    "tavily-python>=0.5.0",
]
